http: harden HTTP/2 fetch client + split into h2_client/
Hardening pass over the experimental `fetch({protocol:"http2"})` client
landed in #29766. Covers memory safety, RFC 9113 conformance, DoS
bounds, and a couple of perf hotspots; also splits the 1.7k-line
H2Client.zig into a `src/http/h2_client/` module.

Memory safety / lifecycle
- Detach + RST before doRedirect when `is_redirect_pending` is set on a
  has-body response, so the old Stream can't keep a dangling
  `*HTTPClient` after the redirect re-dispatches.
- `removeStream()` moves a half-buffered HEADERS fragment into
  `orphan_header_block` when the stream is torn down mid-CONTINUATION,
  so the trailing CONTINUATION decodes against the full block instead
  of desyncing the HPACK dynamic table for siblings.
- `adopt()` re-dispatches when `!hasHeadroom()`; `canPool()` excludes
  `MAX_CONCURRENT_STREAMS=0` so a useless session is never resumed from
  the keep-alive pool (was a debug assert).
- Encoder-poison path bounces `pending_attach` and releases the session
  instead of leaking it until idle-timeout.

DoS bounds
- `local_max_header_list_size` (256 KiB) capped on header-block
  accumulation and on the decoded header list; advertised via
  `SETTINGS_MAX_HEADER_LIST_SIZE`. Stops CONTINUATION-flood / HPACK-
  amplification OOM.
- PING/SETTINGS-ACK reflection capped at 1 MiB queued write_buffer
  (`ENHANCE_YOUR_CALM`).
- First server frame must be SETTINGS (RFC 9113 §3.4); GOAWAY-first no
  longer parks coalesced waiters forever.

Correctness
- DATA after END_STREAM/RST in the same read is rejected, not appended.
- RST_STREAM(NO_ERROR) mid-body fails the request instead of delivering
  a truncated body as success.
- REFUSED_STREAM only retries when no part of the response was
  delivered (`status_code == 0`).
- Trailers HEADERS without END_STREAM rejected; pseudo-headers in
  trailers rejected.
- GOAWAY no longer fails streams ≤ last_stream_id that already
  completed in the same read.
- `data_bytes_received` tracks raw DATA so Content-Length overshoot is
  caught (the body handler clamps `total_body_received`).
- Response field-section validated: uppercase names, hop-by-hop fields,
  pseudo-after-regular, duplicate/unknown pseudo → PROTOCOL_ERROR.
- WINDOW_UPDATE on idle stream → PROTOCOL_ERROR; PRIORITY framing
  validated; TE filtered to "trailers"-only; `failAll` emits GOAWAY.
- Expect:100-continue: drop content-length and RST(NO_ERROR) on
  final-first instead of malformed 0-byte DATA(END_STREAM).
- Detach RSTs whenever the server still sees the stream open, and
  skips RST on fully-closed streams.

Performance
- `parseFrames` reads from the socket slice when no partial frame is
  carried over (one fewer body memcpy).
- 256 KiB `write_buffer` high-water mark in `writeDataWindowed` so a
  large window grant doesn't duplicate the body in memory.
- `drainSendBodies` round-robins one max-frame-size slice per stream.
- `replenishWindow` runs after the deliver loop.
- Idle timer only re-arms on stream-advancing frames (PING-only traffic
  can't keep a stalled upload alive).
- `writeRequest`: header classification via one `ComptimeStringMap`
  lookup; encoded block reuses session-level `encode_scratch`
  (clearAndFree past 64 KiB); names lowercased via stack
  `copyLowercaseIfNeeded`.

Layout
- `H2Client.zig` → 55-line namespace re-exporting:
  `h2_client/{ClientSession,Stream,PendingConnect,dispatch,encode}.zig`

Tests
- New `test/js/web/fetch/fetch-http2-adversarial.test.ts` (raw-frame
  TLS server): CONTINUATION flood bound, MCS=0, repeated WINDOW_UPDATE
  overflow, no-SETTINGS, DATA-after-END_STREAM, tiny-DATA flood,
  unknown frame type, bad padding, RST(NO_ERROR) mid-body, trailers-
  no-END_STREAM, GOAWAY-after-complete, GOAWAY-before-SETTINGS.
